+++ test/Nirum/Targets/PythonSpec.hs
@@ -3,11 +3,15 @@
 module Nirum.Targets.PythonSpec where
 
 import qualified Data.Map.Strict as M
+import Data.Either
 import System.FilePath ((</>))
 import Test.Hspec.Meta
 
+import Nirum.Constructs.Annotation hiding (null)
+import Nirum.Constructs.Annotation.Internal
 import Nirum.Constructs.Module (Module (Module))
-import Nirum.Package.Metadata (Target (compilePackage))
+import Nirum.Constructs.TypeDeclaration hiding (Text)
+import qualified Nirum.Package.Metadata as M
 import Nirum.Targets.Python
     ( Source (Source)
     , parseModulePath
@@ -19,7 +23,7 @@ spec = do
     describe "compilePackage" $ do
         it "returns a Map of file paths and their contents to generate" $ do
             let (Source pkg _) = makeDummySource $ Module [] Nothing
-                files = compilePackage pkg
+                files = M.compilePackage pkg
                 directoryStructure =
                     [ "src-py2" </> "foo" </> "__init__.py"
                     , "src-py2" </> "foo" </> "bar" </> "__init__.py"
@@ -34,7 +38,7 @@ spec = do
         it "creates an emtpy Python package directory if necessary" $ do
             let (Source pkg _) = makeDummySource' ["test"] (Module [] Nothing)
                                                   []
-                files = compilePackage pkg
+                files = M.compilePackage pkg
                 directoryStructure =
                     [ "src-py2" </> "test" </> "__init__.py"
                     , "src-py2" </> "test" </> "foo" </> "__init__.py"
@@ -51,7 +55,7 @@ spec = do
         it "generates renamed package dirs if renames are configured" $ do
             let (Source pkg _) = makeDummySource' [] (Module [] Nothing)
                                                   [(["foo"], ["quz"])]
-                files = compilePackage pkg
+                files = M.compilePackage pkg
                 directoryStructure =
                     [ "src-py2" </> "quz" </> "__init__.py"
                     , "src-py2" </> "quz" </> "bar" </> "__init__.py"
@@ -65,7 +69,7 @@ spec = do
             M.keysSet files `shouldBe` directoryStructure
             let (Source pkg' _) = makeDummySource' [] (Module [] Nothing)
                                                    [(["foo", "bar"], ["bar"])]
-                files' = compilePackage pkg'
+                files' = M.compilePackage pkg'
                 directoryStructure' =
                     [ "src-py2" </> "foo" </> "__init__.py"
                     , "src-py2" </> "bar" </> "__init__.py"
@@ -90,3 +94,43 @@ spec = do
         parseModulePath "foo..bar" `shouldBe` Nothing
         parseModulePath "foo.bar>" `shouldBe` Nothing
         parseModulePath "foo.bar-" `shouldBe` Nothing
+
+    describe "@numeric-constraints" $ do
+        it "fails if unsupported arguments are present" $ do
+            let Right annots = fromList
+                    [ Annotation
+                          "numeric-constraints"
+                          [("min", Integer 1), ("unsupported", Integer 2)]
+                    ]
+            compareErrorMessage annots
+
+        it "fails if unsupported arguments type is given" $ do
+            let Right annots = fromList
+                    [ Annotation
+                          "numeric-constraints"
+                          [("min", Integer 1), ("max", Text "2")]
+                    ]
+            compareErrorMessage annots
+
+        it "success" $ do
+            let Right annots = fromList
+                    [ Annotation
+                          "numeric-constraints"
+                          [("min", Integer 1), ("max", Integer 2)]
+                    ]
+            let Just result = getResult annots
+            isRight result `shouldBe` True
+      where
+          compareErrorMessage annots = do
+              let Just result = getResult annots
+              let Left errorMessage = result
+              errorMessage `shouldBe`
+                 "Unsupported arguments on @numeric-constraints"
+
+          getResult annots =
+              let
+                  unboxed = TypeDeclaration "foo" (UnboxedType "int32") annots
+                  (Source pkg _) = makeDummySource $ Module [unboxed] Nothing
+                  files = M.compilePackage pkg
+              in
+                  M.lookup ("src" </> "foo" </> "__init__.py") files